+++ fakernaija/providers/__init__.py
@@ -1,0 +1,12 @@
+"""The providers subpackage contains various data providers for generating Nigerian-specific fake data.
+
+Each provider class is responsible for generating a specific type of data, such as names, 
+phone numbers, schools, and states. These classes are used by the main Faker class to 
+produce randomized data that is useful for testing and development purposes.
+
+Modules:
+- names: Provides functionality for generating Nigerian names.
+- phonenumbers: Provides functionality for generating Nigerian phone numbers.
+- schools: Provides functionality for generating information about Nigerian educational institutions.
+- states: Provides functionality for generating information about Nigerian states.
+"""