+++ fakernaija/providers/data/__init__.py
@@ -1,0 +1,6 @@
+"""The data subpackage contains static data files used by the various data providers in the fakernaija package.
+
+This data includes names, state details, school details, and other relevant information
+specific to Nigeria. These files are utilized by the provider classes to generate realistic and 
+randomized Nigerian data for testing and development purposes.
+"""